LPS-80145 Move FragmentEntryLink and related components to a separated directory
