+++ modules/apps/fragment/fragment-display-web/src/main/resources/META-INF/resources/js/FragmentEntryDisplay.es.js
@@ -2,7 +2,7 @@ import Component from 'metal-component';
 import {Config} from 'metal-state';
 import Soy from 'metal-soy';
 
-import 'layout-admin-web/js/fragments_editor/FragmentEntryLink.es';
+import 'layout-admin-web/js/fragments_editor/render/FragmentEntryLink.es';
 import templates from './FragmentEntryDisplay.soy';
 
 /**
