+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/render/FragmentEditableFieldTooltip.es.js
